# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SerialCuts"
  },
  "paths": {
...
    "/api/episodes/{episode_id}/probe": {
      "post": {
        "summary": "Probe Episode",
        "operationId": "probe_episode_api_episodes__episode_id__probe_post",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}": {
      "delete": {
        "summary": "Delete Episode Endpoint",
        "operationId": "delete_episode_endpoint_api_episodes__episode_id__delete",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/enqueue": {
      "post": {
        "summary": "Enqueue Episode",
        "operationId": "enqueue_episode_api_episodes__episode_id__enqueue_post",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EnqueueEpisodeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/seasons/{season_id}/enqueue": {
      "post": {
        "summary": "Enqueue Season",
        "operationId": "enqueue_season_api_seasons__season_id__enqueue_post",
        "parameters": [
          {
            "name": "season_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueSeasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobRead"
                  },
                  "title": "Response Enqueue Season Api Seasons  Season Id  Enqueue Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/stage2": {
      "post": {
        "summary": "Run Stage2 Episode",
        "operationId": "run_stage2_episode_api_episodes__episode_id__stage2_post",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stage2RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/stage3": {
      "post": {
        "summary": "Run Stage3 Episode",
        "operationId": "run_stage3_episode_api_episodes__episode_id__stage3_post",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stage3RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/candidates": {
      "get": {
        "summary": "List Episode Candidates",
        "operationId": "list_episode_candidates_api_episodes__episode_id__candidates_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRead"
                  },
                  "title": "Response List Episode Candidates Api Episodes  Episode Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/quality": {
      "get": {
        "summary": "Episode Quality",
        "operationId": "episode_quality_api_episodes__episode_id__quality_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeQualityRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/proxy": {
      "get": {
        "summary": "Episode Proxy",
        "operationId": "episode_proxy_api_episodes__episode_id__proxy_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "start_time": {
            "type": "number",
            "title": "Start Time"
          },
          "end_time": {
            "type": "number",
            "title": "End Time"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "moment_type": {
            "type": "string",
            "title": "Moment Type"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "scores_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scores Json"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "problems_json": {
            "items": {},
            "type": "array",
            "title": "Problems Json"
          },
          "crop_mode": {
            "type": "string",
            "title": "Crop Mode"
          },
          "crop_offset_x": {
            "type": "number",
            "title": "Crop Offset X"
          },
          "crop_scale": {
            "type": "number",
            "title": "Crop Scale"
          },
          "crop_keyframes_json": {
            "items": {},
            "type": "array",
            "title": "Crop Keyframes Json"
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Path"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "story_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Order"
          },
          "story_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Role"
          },
          "continuity_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuity Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "episode_id",
          "start_time",
          "end_time",
          "title",
          "description",
          "moment_type",
          "score",
          "scores_json",
          "rationale",
          "problems_json",
          "crop_mode",
          "crop_offset_x",
          "crop_scale",
          "crop_keyframes_json",
          "thumbnail_path",
          "status"
        ],
        "title": "CandidateRead"
      },
...
      "EnqueueEpisodeRequest": {
        "properties": {
          "resume_from_stage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(stage2_media|stage3_candidates|auto_export)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume From Stage"
          },
          "auto": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "max_clips": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Clips"
          },
          "use_nvenc": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Nvenc"
          }
        },
        "type": "object",
        "title": "EnqueueEpisodeRequest"
      },
      "EnqueueSeasonRequest": {
        "properties": {
          "auto": {
            "type": "boolean",
            "title": "Auto",
            "default": false
          }
        },
        "type": "object",
        "title": "EnqueueSeasonRequest"
      },
...
      "EpisodeQualityRead": {
        "properties": {
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "transcript_segments": {
            "type": "integer",
            "title": "Transcript Segments"
          },
          "words": {
            "type": "integer",
            "title": "Words"
          },
          "scenes": {
            "type": "integer",
            "title": "Scenes"
          },
          "candidates": {
            "type": "integer",
            "title": "Candidates"
          },
          "approved": {
            "type": "integer",
            "title": "Approved"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "rendered": {
            "type": "integer",
            "title": "Rendered"
          },
          "average_score": {
            "type": "integer",
            "title": "Average Score"
          },
          "problem_candidates": {
            "type": "integer",
            "title": "Problem Candidates"
          },
          "top_problems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Problems"
          }
        },
        "type": "object",
        "required": [
          "episode_id",
          "stage",
          "transcript_segments",
          "words",
          "scenes",
          "candidates",
          "approved",
          "rejected",
          "rendered",
          "average_score",
          "problem_candidates",
          "top_problems"
        ],
        "title": "EpisodeQualityRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "episode_id",
          "kind",
          "status",
          "current_stage",
          "progress",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
...
      "Stage2RunResponse": {
        "properties": {
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "audio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Path"
          },
          "proxy_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Path"
          },
          "transcript_segments": {
            "type": "integer",
            "title": "Transcript Segments"
          },
          "scenes": {
            "type": "integer",
            "title": "Scenes"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "episode_id",
          "stage",
          "audio_path",
          "proxy_path",
          "transcript_segments",
          "scenes"
        ],
        "title": "Stage2RunResponse"
      },
      "Stage3RunResponse": {
        "properties": {
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "outline_created": {
            "type": "boolean",
            "title": "Outline Created"
          },
          "candidates": {
            "type": "integer",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "episode_id",
          "stage",
          "outline_created",
          "candidates"
        ],
        "title": "Stage3RunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}